{"schema":"decodex/commit/1","summary":"Split knowledge page readback DTO modules","authority":"manual"}
